The app in its current state is fully functional, however, it requires that the user manually inputs latitude and longitude values to specify location.
Next step is to add Google Maps integration to allow for the user to drop a
pin to specify location.

Might also want to add sorting options, i.e. to allow the user to find the
most recently uploaded video or the closest video to their specified location,
instead of just the most viewed video.